@@ -120,19 +120,19 @@
     "*.{css,json}": "biome format --write"
   },
   "devDependencies": {
-    "@biomejs/biome": "^2.2.5",
+    "@biomejs/biome": "^2.2.7",
     "@commitlint/cli": "^20.1.0",
     "@commitlint/config-conventional": "^20.0.0",
     "@electron/fuses": "^2.0.0",
     "@eslint/compat": "^1.4.0",
     "@eslint/eslintrc": "^3.3.1",
-    "@eslint/js": "^9.37.0",
-    "@playwright/test": "1.56.0",
+    "@eslint/js": "^9.38.0",
+    "@playwright/test": "1.56.1",
     "@podman-desktop/api": "workspace:",
-    "@rollup/plugin-commonjs": "^28.0.6",
+    "@rollup/plugin-commonjs": "^28.0.8",
     "@rollup/plugin-json": "^6.1.0",
     "@rollup/plugin-node-resolve": "^16.0.3",
-    "@rollup/plugin-typescript": "^12.1.4",
+    "@rollup/plugin-typescript": "^12.2.0",
     "@tsconfig/strictest": "^2.0.6",
     "@types/adm-zip": "^0.5.7",
     "@types/dockerode": "^3.3.44",
@@ -147,13 +147,13 @@
     "@typescript-eslint/eslint-plugin": "^8.46.0",
     "@typescript-eslint/parser": "^8.46.0",
     "@vitest/coverage-v8": "^3.2.4",
-    "@vitest/eslint-plugin": "^1.3.17",
+    "@vitest/eslint-plugin": "^1.3.23",
     "cross-env": "10.1.0",
     "dts-for-context-bridge": "0.7.1",
-    "electron": "38.2.2",
+    "electron": "38.4.0",
     "electron-builder": "26.0.12",
     "electron-builder-notarize": "^1.5.2",
-    "eslint": "^9.37.0",
+    "eslint": "^9.38.0",
     "eslint-import-resolver-custom-alias": "^1.3.2",
     "eslint-import-resolver-typescript": "^4.4.4",
     "eslint-plugin-etc": "^2.0.3",
@@ -165,20 +165,20 @@
     "eslint-plugin-sonarjs": "^3.0.5",
     "eslint-plugin-unicorn": "^61.0.2",
     "fzstd": "^0.1.1",
-    "get-tsconfig": "^4.12.0",
+    "get-tsconfig": "^4.13.0",
     "globals": "^16.4.0",
     "husky": "^9.1.7",
-    "inversify": "^7.10.2",
+    "inversify": "^7.10.3",
     "js-yaml": "^4.1.0",
-    "lint-staged": "^16.2.4",
+    "lint-staged": "^16.2.6",
     "minimist": "^1.2.8",
-    "msw": "^2.11.5",
+    "msw": "^2.11.6",
     "prettier": "^3.6.2",
     "proxy": "^2.2.0",
     "typescript": "5.9.3",
     "typescript-eslint": "^8.46.0",
     "validator": "^13.15.15",
-    "vite": "^7.1.9",
+    "vite": "^7.1.11",
     "vitest": "^3.2.4",
     "xvfb-maybe": "^0.2.1"
   },
