@@ -17,7 +17,7 @@
     "@fortawesome/free-regular-svg-icons": "^7.1.0",
     "@fortawesome/free-solid-svg-icons": "^7.1.0",
     "@sveltejs/vite-plugin-svelte": "6.2.1",
-    "@tailwindcss/vite": "^4.1.14",
+    "@tailwindcss/vite": "^4.1.15",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/svelte": "^5.2.8",
@@ -27,7 +27,7 @@
     "@types/validator": "^13.15.3",
     "@typescript-eslint/eslint-plugin": "^8.46.0",
     "@xterm/addon-serialize": "^0.13.0",
-    "eslint-plugin-svelte": "^3.12.4",
+    "eslint-plugin-svelte": "^3.12.5",
     "filesize": "^11.0.13",
     "humanize-duration": "^3.33.1",
     "jsdom": "^27.0.0-beta.3",
@@ -35,15 +35,15 @@
     "micromark-extension-directive": "^4.0.0",
     "moment": "^2.30.1",
     "monaco-editor": "^0.54.0",
-    "svelte": "5.39.11",
+    "svelte": "5.41.2",
     "svelte-check": "^4.3.3",
-    "svelte-eslint-parser": "^1.3.3",
+    "svelte-eslint-parser": "^1.4.0",
     "svelte-fa": "^4.0.4",
     "svelte-steps": "2.4.1",
-    "tailwindcss": "^4.1.14",
+    "tailwindcss": "^4.1.15",
     "tslib": "^2.8.1",
     "typescript": "^5.9.3",
-    "vite": "^7.1.9",
+    "vite": "^7.1.11",
     "vitest": "^3.2.4",
     "@xterm/xterm": "^5.5.0",
     "@xterm/addon-fit": "^0.10.0",
