@@ -149,6 +149,7 @@
     "svelte-fa": "^4.0.0"
   },
   "dependencies": {
+    "@floating-ui/dom": "^1.7.4",
     "@fortawesome/fontawesome-free": "^7.1.0",
     "@fortawesome/free-brands-svg-icons": "^7.1.0",
     "@fortawesome/free-regular-svg-icons": "^7.1.0",
@@ -160,22 +161,22 @@
   "devDependencies": {
     "@sveltejs/package": "^2.5.4",
     "@sveltejs/vite-plugin-svelte": "6.2.1",
-    "@tailwindcss/vite": "^4.1.14",
+    "@tailwindcss/vite": "^4.1.15",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/svelte": "^5.2.8",
     "@testing-library/user-event": "^14.6.1",
     "@tsconfig/svelte": "^5.0.5",
     "@typescript-eslint/eslint-plugin": "^8.46.0",
-    "eslint-plugin-svelte": "^3.12.4",
+    "eslint-plugin-svelte": "^3.12.5",
     "jsdom": "^27.0.0-beta.3",
-    "svelte": "5.39.11",
+    "svelte": "5.41.2",
     "svelte-check": "^4.3.3",
-    "svelte-eslint-parser": "^1.3.3",
-    "tailwindcss": "^4.1.14",
+    "svelte-eslint-parser": "^1.4.0",
+    "tailwindcss": "^4.1.15",
     "tslib": "^2.8.1",
     "typescript": "^5.9.3",
-    "vite": "^7.1.9",
+    "vite": "^7.1.11",
     "vitest": "^3.2.4"
   }
 }
