@@ -17,33 +17,33 @@
     "@fortawesome/free-regular-svg-icons": "^7.1.0",
     "@fortawesome/free-solid-svg-icons": "^7.1.0",
     "@podman-desktop/ui-svelte": "workspace:*",
-    "eslint": "^9.37.0",
+    "eslint": "^9.38.0",
     "svelte-fa": "^4.0.4"
   },
   "devDependencies": {
-    "@storybook/addon-docs": "^9.1.10",
-    "@storybook/addon-links": "^9.1.10",
+    "@storybook/addon-docs": "^9.1.13",
+    "@storybook/addon-links": "^9.1.13",
     "@storybook/addon-svelte-csf": "5.0.9",
-    "@storybook/svelte-vite": "^9.1.10",
+    "@storybook/svelte-vite": "^9.1.13",
     "@sveltejs/package": "^2.5.4",
     "@sveltejs/vite-plugin-svelte": "6.2.1",
-    "@tailwindcss/vite": "^4.1.14",
+    "@tailwindcss/vite": "^4.1.15",
     "@tsconfig/svelte": "^5.0.5",
     "@typescript-eslint/eslint-plugin": "^8.46.0",
     "autoprefixer": "^10.4.21",
     "cross-env": "10.1.0",
-    "eslint-plugin-storybook": "9.1.10",
-    "eslint-plugin-svelte": "^3.12.4",
+    "eslint-plugin-storybook": "9.1.13",
+    "eslint-plugin-svelte": "^3.12.5",
     "postcss": "^8.5.6",
     "postcss-load-config": "^6.0.1",
     "react": "18.2.0",
     "react-dom": "18.3.1",
     "storybook": "^9.1.10",
-    "svelte": "5.39.11",
+    "svelte": "5.41.2",
     "svelte-check": "^4.3.3",
-    "tailwindcss": "^4.1.14",
+    "tailwindcss": "^4.1.15",
     "typescript": "^5.9.3",
-    "vite": "^7.1.9",
+    "vite": "^7.1.11",
     "vitest": "^3.2.4"
   }
 }
