@@ -22,9 +22,9 @@
     "publish:next": "pnpm publish --registry=https://registry.npmjs.org/ --no-git-tag-version --new-version 0.0.1-\"$(date +%s)\""
   },
   "devDependencies": {
-    "@playwright/test": "1.56.0",
+    "@playwright/test": "1.56.1",
     "@types/node": "^22",
-    "electron": "38.2.2",
+    "electron": "38.4.0",
     "typescript": "^5.9.3",
     "vitest": "^3.2.4"
   }
