@@ -27,6 +27,20 @@ import { Disposable, provider as apiProvider } from '@podman-desktop/api';
 import type { Mock } from 'vitest';
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
 
+import {
+  CLEANUP_REQUIRED_MACHINE_KEY,
+  CREATE_WSL_MACHINE_OPTION_SELECTED_KEY,
+  PODMAN_MACHINE_CPU_SUPPORTED_KEY,
+  PODMAN_MACHINE_DISK_SUPPORTED_KEY,
+  PODMAN_MACHINE_EDIT_CPU,
+  PODMAN_MACHINE_EDIT_DISK_SIZE,
+  PODMAN_MACHINE_EDIT_MEMORY,
+  PODMAN_MACHINE_EDIT_ROOTFUL,
+  PODMAN_MACHINE_MEMORY_SUPPORTED_KEY,
+  WSL_HYPERV_ENABLED_KEY,
+} from '/@/constants';
+import type { ConnectionJSON, MachineInfo, MachineJSON } from '/@/types';
+
 import * as extension from './extension';
 import {
   initCheckAndRegisterUpdate,
@@ -95,7 +109,7 @@ const providerWithStoppedStatus = {
   status: 'stopped' as extensionApi.ProviderStatus,
 };
 
-const machineInfo: extension.MachineInfo = {
+const machineInfo: MachineInfo = {
   cpus: 1,
   diskSize: 1000000,
   memory: 10000000,
@@ -122,7 +136,7 @@ const machineDefaultName = 'podman-machine-default';
 const machine1Name = 'podman-machine-1';
 
 // Create fake of MachineJSON
-let fakeMachineJSON: extension.MachineJSON[];
+let fakeMachineJSON: MachineJSON[];
 let fakeMachineInfoJSON: unknown;
 
 const telemetryLogger: extensionApi.TelemetryLogger = {
@@ -820,7 +834,7 @@ test('test checkDefaultMachine, if the machine running is not default, the funct
 
 test('checkDefaultMachine: do not prompt if the running machine is already the default', async () => {
   // Create fake of MachineJSON
-  const fakeJSON: extension.MachineJSON[] = [
+  const fakeJSON: MachineJSON[] = [
     {
       Name: 'podman-machine-default',
       CPUs: 2,
@@ -849,7 +863,7 @@ test('checkDefaultMachine: do not prompt if the running machine is already the d
     },
   ];
 
-  const fakeConnectionJSON: extension.ConnectionJSON[] = [
+  const fakeConnectionJSON: ConnectionJSON[] = [
     {
       Name: machineDefaultName,
       URI: 'uri',
@@ -982,10 +996,10 @@ test('if a machine failed to start with a wsl distro not found error but the ski
 
 test('test checkDefaultMachine - if there is no machine marked as default, take the default system connection to retrieve it. Prompt as it is not running', async () => {
   // Create fake of MachineJSON
-  const fakeJSON: extension.MachineJSON[] = fakeMachineJSON;
+  const fakeJSON: MachineJSON[] = fakeMachineJSON;
   fakeJSON[1].Default = false;
 
-  const fakeConnectionJSON: extension.ConnectionJSON[] = [
+  const fakeConnectionJSON: ConnectionJSON[] = [
     {
       Name: machineDefaultName,
       URI: 'uri',
@@ -1032,10 +1046,10 @@ test('test checkDefaultMachine - if there is no machine marked as default, take 
 
 test('test checkDefaultMachine - if there is no machine marked as default, take the default system connection to retrieve it. Do not prompt as it is running', async () => {
   // Create fake of MachineJSON
-  const fakeJSON: extension.MachineJSON[] = fakeMachineJSON;
+  const fakeJSON: MachineJSON[] = fakeMachineJSON;
   fakeJSON[1].Default = false;
 
-  const fakeConnectionJSON: extension.ConnectionJSON[] = [
+  const fakeConnectionJSON: ConnectionJSON[] = [
     {
       Name: machineDefaultName,
       URI: 'uri',
@@ -1208,7 +1222,7 @@ test('test checkDefaultMachine - if user wants to change machine, check that it 
 });
 
 test('test checkDefaultMachine, if the default connection is not in sync with the default machine, the function will prompt', async () => {
-  const fakeConnectionJSON: extension.ConnectionJSON[] = [
+  const fakeConnectionJSON: ConnectionJSON[] = [
     {
       Name: machineDefaultName,
       URI: 'uri',
@@ -1552,10 +1566,10 @@ test('provider is registered with edit capabilities on MacOS', async () => {
   expect(registeredConnection).toBeDefined();
   expect(registeredConnection?.lifecycle).toBeDefined();
   expect(registeredConnection?.lifecycle?.edit).toBeDefined();
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_CPU, true);
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_MEMORY, true);
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_DISK_SIZE, true);
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_ROOTFUL, true);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_CPU, true);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_MEMORY, true);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_DISK_SIZE, true);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_ROOTFUL, true);
 });
 
 test('display name is beautified version of the name', async () => {
@@ -1625,10 +1639,10 @@ test('provider is registered with limited edit capabilities on (non-HyperV) Wind
   expect(registeredConnection).toBeDefined();
   expect(registeredConnection?.lifecycle).toBeDefined();
   expect(registeredConnection?.lifecycle?.edit).toBeDefined();
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_CPU, false);
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_MEMORY, false);
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_DISK_SIZE, false);
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_ROOTFUL, true);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_CPU, false);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_MEMORY, false);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_DISK_SIZE, false);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_ROOTFUL, true);
 });
 
 test('provider is registered with limited edit capabilities on (HyperV) Windows', async () => {
@@ -1668,10 +1682,10 @@ test('provider is registered with limited edit capabilities on (HyperV) Windows'
   expect(registeredConnection).toBeDefined();
   expect(registeredConnection?.lifecycle).toBeDefined();
   expect(registeredConnection?.lifecycle?.edit).toBeDefined();
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_CPU, true);
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_MEMORY, true);
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_DISK_SIZE, false);
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_ROOTFUL, true);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_CPU, true);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_MEMORY, true);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_DISK_SIZE, false);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_ROOTFUL, true);
 
   await registeredConnection?.lifecycle?.edit?.({} as unknown as extensionApi.LifecycleContext, {
     'podman.machine.cpus': 1,
@@ -1713,9 +1727,9 @@ test('provider is registered without edit capabilities on Linux', async () => {
   expect(registeredConnection).toBeDefined();
   expect(registeredConnection?.lifecycle).toBeDefined();
   expect(registeredConnection?.lifecycle?.edit).toBeUndefined();
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_CPU, false);
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_MEMORY, false);
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_EDIT_DISK_SIZE, false);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_CPU, false);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_MEMORY, false);
+  expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_EDIT_DISK_SIZE, false);
 });
 
 test('Even with getJSONMachineList erroring, do not show setup notification on Linux', async () => {
@@ -1753,7 +1767,7 @@ test('Should notify clean machine if getJSONMachineList is erroring due to an in
   });
   await expect(extension.updateMachines(provider, podmanConfiguration)).rejects.toThrow('description');
   expect(extensionApi.window.showNotification).toBeCalled();
-  expect(extensionApi.context.setValue).toBeCalledWith(extension.CLEANUP_REQUIRED_MACHINE_KEY, true);
+  expect(extensionApi.context.setValue).toBeCalledWith(CLEANUP_REQUIRED_MACHINE_KEY, true);
 });
 
 test('No updates of machines in parallel', async () => {
@@ -2438,9 +2452,9 @@ describe('calcPodmanMachineSetting', () => {
   test('setValue to true if OS is MacOS', async () => {
     vi.mocked(extensionApi.env).isWindows = false;
     await extension.calcPodmanMachineSetting();
-    expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_CPU_SUPPORTED_KEY, true);
-    expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_MEMORY_SUPPORTED_KEY, true);
-    expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_DISK_SUPPORTED_KEY, true);
+    expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_CPU_SUPPORTED_KEY, true);
+    expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_MEMORY_SUPPORTED_KEY, true);
+    expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_DISK_SUPPORTED_KEY, true);
   });
   test('setValue to true if OS is Windows and uses HyperV', async () => {
     vi.mocked(extensionApi.env).isWindows = true;
@@ -2459,18 +2473,18 @@ describe('calcPodmanMachineSetting', () => {
       });
     });
     await extension.calcPodmanMachineSetting();
-    expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_CPU_SUPPORTED_KEY, true);
-    expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_MEMORY_SUPPORTED_KEY, true);
-    expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_DISK_SUPPORTED_KEY, true);
+    expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_CPU_SUPPORTED_KEY, true);
+    expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_MEMORY_SUPPORTED_KEY, true);
+    expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_DISK_SUPPORTED_KEY, true);
   });
   test('setValue to true if OS is Windows and uses WSL', async () => {
     vi.mocked(extensionApi.env).isWindows = true;
     process.env.CONTAINERS_MACHINE_PROVIDER = 'wsl';
     vi.spyOn(podmanConfiguration, 'matchRegexpInContainersConfig').mockResolvedValue(false);
     await extension.calcPodmanMachineSetting();
-    expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_CPU_SUPPORTED_KEY, false);
-    expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_MEMORY_SUPPORTED_KEY, false);
-    expect(extensionApi.context.setValue).toBeCalledWith(extension.PODMAN_MACHINE_DISK_SUPPORTED_KEY, false);
+    expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_CPU_SUPPORTED_KEY, false);
+    expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_MEMORY_SUPPORTED_KEY, false);
+    expect(extensionApi.context.setValue).toBeCalledWith(PODMAN_MACHINE_DISK_SUPPORTED_KEY, false);
   });
 });
 
@@ -2965,7 +2979,7 @@ test('getJSONMachineList should only get machines from wsl if hyperv is not enab
       }
     });
   });
-  const fakeJSON: extension.MachineJSON[] = [
+  const fakeJSON: MachineJSON[] = [
     {
       Name: 'podman-machine-default',
       CPUs: 2,
@@ -3030,7 +3044,7 @@ test('getJSONMachineList should only get machines from hyperv if wsl is not enab
       }
     });
   });
-  const fakeJSON: extension.MachineJSON[] = [
+  const fakeJSON: MachineJSON[] = [
     {
       Name: 'podman-machine-default',
       CPUs: 2,
@@ -3096,7 +3110,7 @@ test('getJSONMachineList should get machines from hyperv and wsl if both are ena
       }
     });
   });
-  const fakeJSON: extension.MachineJSON[] = [
+  const fakeJSON: MachineJSON[] = [
     {
       Name: 'podman-machine-default',
       CPUs: 2,
@@ -3141,7 +3155,7 @@ describe('updateWSLHyperVEnabledValue', () => {
   });
   test('setValue should be called if new value is different than wslAndHypervEnabled', async () => {
     extension.updateWSLHyperVEnabledContextValue(false);
-    expect(extensionApi.context.setValue).toBeCalledWith(extension.WSL_HYPERV_ENABLED_KEY, false);
+    expect(extensionApi.context.setValue).toBeCalledWith(WSL_HYPERV_ENABLED_KEY, false);
   });
   test('setValue should not be called if new value is equal to wslAndHypervEnabled', async () => {
     extension.updateWSLHyperVEnabledContextValue(true);
@@ -3160,10 +3174,7 @@ describe('connectionAuditor', () => {
     await extension.connectionAuditor({
       'podman.factory.machine.win.provider': 'wsl',
     });
-    expect(extensionApi.context.setValue).toHaveBeenLastCalledWith(
-      extension.CREATE_WSL_MACHINE_OPTION_SELECTED_KEY,
-      true,
-    );
+    expect(extensionApi.context.setValue).toHaveBeenLastCalledWith(CREATE_WSL_MACHINE_OPTION_SELECTED_KEY, true);
   });
   test('check if podman.isCreateWSLOptionSelected is set to true if podman.factory.machine.win.provider is undefined but wsl is enabled', async () => {
     // be sure isCreateWSLOptionSelected is set to false
@@ -3177,10 +3188,7 @@ describe('connectionAuditor', () => {
     await extension.connectionAuditor({
       'podman.factory.machine.win.provider': undefined,
     });
-    expect(extensionApi.context.setValue).toHaveBeenLastCalledWith(
-      extension.CREATE_WSL_MACHINE_OPTION_SELECTED_KEY,
-      true,
-    );
+    expect(extensionApi.context.setValue).toHaveBeenLastCalledWith(CREATE_WSL_MACHINE_OPTION_SELECTED_KEY, true);
   });
   test('check if podman.isCreateWSLOptionSelected is set to false if podman.factory.machine.win.provider = hyperv', async () => {
     // be sure isCreateWSLOptionSelected is set to true
@@ -3192,10 +3200,7 @@ describe('connectionAuditor', () => {
     await extension.connectionAuditor({
       'podman.factory.machine.win.provider': 'hyperv',
     });
-    expect(extensionApi.context.setValue).toHaveBeenLastCalledWith(
-      extension.CREATE_WSL_MACHINE_OPTION_SELECTED_KEY,
-      false,
-    );
+    expect(extensionApi.context.setValue).toHaveBeenLastCalledWith(CREATE_WSL_MACHINE_OPTION_SELECTED_KEY, false);
   });
   test('check if podman.isCreateWSLOptionSelected is set to false if podman.factory.machine.win.provider is undefined and wsl is NOT enabled', async () => {
     // be sure isCreateWSLOptionSelected is set to true
@@ -3209,10 +3214,7 @@ describe('connectionAuditor', () => {
     await extension.connectionAuditor({
       'podman.factory.machine.win.provider': undefined,
     });
-    expect(extensionApi.context.setValue).toHaveBeenLastCalledWith(
-      extension.CREATE_WSL_MACHINE_OPTION_SELECTED_KEY,
-      false,
-    );
+    expect(extensionApi.context.setValue).toHaveBeenLastCalledWith(CREATE_WSL_MACHINE_OPTION_SELECTED_KEY, false);
   });
 });
 
