@@ -25,6 +25,7 @@ import { EventEmitter } from 'stream-json/Assembler.js';
 import { afterEach, beforeAll, beforeEach, describe, expect, test, vi } from 'vitest';
 
 import type { ContributionInfo } from '/@api/contribution-info.js';
+import type { IDisposable } from '/@api/disposable.js';
 
 import * as util from '../util.js';
 import type { ApiSenderType } from './api.js';
@@ -33,7 +34,6 @@ import type { ComposeObject, DockerExtensionMetadata } from './contribution-mana
 import { ContributionManager } from './contribution-manager.js';
 import type { Directories } from './directories.js';
 import type { Proxy } from './proxy.js';
-import type { IDisposable } from './types/disposable.js';
 import { Exec } from './util/exec.js';
 
 let contributionManager: TestContributionManager;
