@@ -18,10 +18,9 @@
 
 /* eslint-disable @typescript-eslint/no-unsafe-function-type */
 
+import type { IDisposable } from '/@api/disposable.js';
 import type { DisposableGroup } from '/@api/disposable-group.js';
 import type { Event } from '/@api/event.js';
-
-import type { IDisposable } from '../types/disposable.js';
 
 type Callback = (...args: unknown[]) => unknown;
 class CallbackList implements Iterable<Callback> {
