@@ -20,11 +20,11 @@ import type { FileSystemWatcher, Uri } from '@podman-desktop/api';
 import type { FileMatcher } from 'get-tsconfig';
 import { inject, injectable } from 'inversify';
 
+import { IDisposable } from '/@api/disposable.js';
 import type { Event } from '/@api/event.js';
 
 import { Emitter } from '../events/emitter.js';
 import { FilesystemMonitoring } from '../filesystem-monitoring.js';
-import type { IDisposable } from '../types/disposable.js';
 import type { AnalyzedExtension } from './extension-analyzer.js';
 import type { ActivatedExtension } from './extension-loader.js';
 import { ExtensionTypeScriptConfigParser } from './extension-tsconfig-parser.js';
