@@ -18,10 +18,10 @@
 
 import { inject, injectable } from 'inversify';
 
+import { IDisposable } from '/@api/disposable.js';
 import { STATUS_BAR_UPDATED_EVENT_NAME, type StatusBarEntry, type StatusBarEntryDescriptor } from '/@api/status-bar.js';
 
 import { ApiSenderType } from '../api.js';
-import type { IDisposable } from '../types/disposable.js';
 
 @injectable()
 export class StatusBarRegistry implements IDisposable {
