@@ -18,7 +18,7 @@
 
 import '@testing-library/jest-dom/vitest';
 
-import { render, screen } from '@testing-library/svelte';
+import { fireEvent, render, screen } from '@testing-library/svelte';
 import { beforeEach, expect, test } from 'vitest';
 
 import ExtensionBadge from './ExtensionBadge.svelte';
@@ -34,10 +34,14 @@ test('Expect to have badge for dd Extension', async () => {
     devMode: false,
   };
   render(ExtensionBadge, { extension });
-  // expect 'Docker Desktop extension' badge
-  const labels = screen.getAllByText('Docker Desktop extension');
 
-  // 2 items
+  const visibleLabel = screen.getByText('Docker Desktop extension');
+  expect(visibleLabel).toBeInTheDocument();
+
+  const tooltipTrigger = screen.getByTestId('tooltip-trigger');
+  await fireEvent.mouseEnter(tooltipTrigger);
+
+  const labels = await screen.findAllByText('Docker Desktop extension');
   expect(labels).toHaveLength(2);
   expect(labels[0]).toBeInTheDocument();
   expect(labels[1]).toBeInTheDocument();
@@ -50,10 +54,14 @@ test('Expect to have badge for pd  built-in Extension', async () => {
     devMode: false,
   };
   render(ExtensionBadge, { extension });
-  // expect 'built-in' badge
-  const labels = screen.getAllByText('built-in Extension');
 
-  // 2 items
+  const visibleLabel = screen.getByText('built-in Extension');
+  expect(visibleLabel).toBeInTheDocument();
+
+  const tooltipTrigger = screen.getByTestId('tooltip-trigger');
+  await fireEvent.mouseEnter(tooltipTrigger);
+
+  const labels = await screen.findAllByText('built-in Extension');
   expect(labels).toHaveLength(2);
   expect(labels[0]).toBeInTheDocument();
   expect(labels[1]).toBeInTheDocument();
@@ -66,10 +74,14 @@ test('Expect to have badge for devMode Extension', async () => {
     devMode: true,
   };
   render(ExtensionBadge, { extension });
-  // expect 'devMode' badge
-  const labels = screen.getAllByText('In Development Mode Extension');
 
-  // 2 items
-  expect(labels).toHaveLength(1);
-  expect(labels[0]).toBeInTheDocument();
+  const visibleLabel = screen.getByText('devMode Extension');
+  expect(visibleLabel).toBeInTheDocument();
+
+  const tooltipTrigger = screen.getByTestId('tooltip-trigger');
+  expect(tooltipTrigger).toBeInTheDocument();
+  await fireEvent.mouseEnter(tooltipTrigger);
+
+  const tooltip = await screen.findByText('In Development Mode Extension');
+  expect(tooltip).toBeInTheDocument();
 });
