@@ -19,7 +19,7 @@
 import '@testing-library/jest-dom/vitest';
 
 import type { ContainerInfo, Port } from '@podman-desktop/api';
-import { render, screen } from '@testing-library/svelte';
+import { fireEvent, render, screen } from '@testing-library/svelte';
 import { afterEach, beforeAll, beforeEach, expect, test, vi } from 'vitest';
 
 import PodColumnActions from './PodColumnActions.svelte';
@@ -83,6 +83,9 @@ test('Expect error message', async () => {
 
   render(PodColumnActions, { object: pod });
 
-  const error = screen.getByText('Pod failed');
+  const tooltipTrigger = screen.getByTestId('tooltip-trigger');
+  await fireEvent.mouseEnter(tooltipTrigger);
+
+  const error = await screen.findByText('Pod failed');
   expect(error).toBeInTheDocument();
 });
