@@ -20,7 +20,7 @@
 
 import '@testing-library/jest-dom/vitest';
 
-import { render, screen, waitFor } from '@testing-library/svelte';
+import { fireEvent, render, screen, waitFor } from '@testing-library/svelte';
 import { tick } from 'svelte';
 import { readable, type Writable, writable } from 'svelte/store';
 import { beforeEach, describe, expect, test, vi } from 'vitest';
@@ -184,6 +184,11 @@ describe.each<{
 
     await tick();
 
+    const tooltipTrigger = screen.queryByTestId('tooltip-trigger');
+    if (tooltipTrigger) {
+      await fireEvent.mouseEnter(tooltipTrigger);
+    }
+
     await vi.waitFor(() => {
       const tooltip = screen.queryByLabelText('tooltip');
       expect(tooltip).toBeNull();
@@ -193,6 +198,13 @@ describe.each<{
   test('expect tooltip when error', async () => {
     setState({ reachable: false, error: 'error message' });
     render(KubernetesCurrentContextConnectionBadge);
+
+    await vi.waitFor(() => {
+      expect(screen.getByRole('status')).toBeInTheDocument();
+    });
+
+    const tooltipTrigger = screen.getByTestId('tooltip-trigger');
+    await fireEvent.mouseEnter(tooltipTrigger);
 
     await vi.waitFor(() => {
       const tooltip = screen.getByLabelText('tooltip');
@@ -208,6 +220,13 @@ describe.each<{
     async () => {
       setState({ reachable: true, offline: true });
       render(KubernetesCurrentContextConnectionBadge);
+
+      await vi.waitFor(() => {
+        expect(screen.getByRole('status')).toBeInTheDocument();
+      });
+
+      const tooltipTrigger = screen.getByTestId('tooltip-trigger');
+      await fireEvent.mouseEnter(tooltipTrigger);
 
       await vi.waitFor(() => {
         const tooltip = screen.getByLabelText('tooltip');
