@@ -58,6 +58,11 @@ test.afterAll(async ({ page, runner }) => {
   }
 });
 
+test.skip(
+  !!isCI && !!isMac,
+  'Tests suite should not run on Mac platform in CICD due to being unable  to ensure Compose is installed',
+);
+
 test.describe.serial('Compose compose workflow verification', { tag: '@smoke' }, () => {
   test.beforeEach(async () => {
     if (cliToolsPage.wasRateLimitReached()) {
@@ -68,7 +73,6 @@ test.describe.serial('Compose compose workflow verification', { tag: '@smoke' },
 
   test('Verify Compose was installed', async ({ page, navigationBar }) => {
     test.skip(!!isCI && isLinux, 'This test should not run on Ubuntu platform in Github Actions');
-    test.skip(!!isMac, 'Currently there is an issue with running this test on macOS platform');
 
     await navigationBar.openSettings();
     const settingsBar = new SettingsBar(page);
@@ -78,7 +82,7 @@ test.describe.serial('Compose compose workflow verification', { tag: '@smoke' },
     const composeBox = new ResourceCliCardPage(page, RESOURCE_NAME);
     const setupButton = composeBox.setupButton;
 
-    if ((await setupButton.count()) > 0) {
+    if ((await setupButton.count()) > 0 && !isMac) {
       await settingsBar.cliToolsTab.click();
       await playExpect(cliToolsPage.toolsTable).toBeVisible({ timeout: 10_000 });
       await playExpect.poll(async () => await cliToolsPage.toolsTable.count()).toBeGreaterThan(0);
